Renewable concern and apply to indirect verifications
