+++ decidim-core/app/permissions/decidim/permissions.rb
@@ -72,10 +72,10 @@ module Decidim
       case permission_action.action
       when :create
         toggle_allow(authorization.user == user && not_already_active?(authorization))
-      when :update
+      when :update, :destroy
         toggle_allow(authorization.user == user && !authorization.granted?)
-      when :destroy
-        toggle_allow(authorization.user == user && !authorization.granted?)
+      when :renew
+        toggle_allow(authorization.user == user && authorization.granted?)
       end
     end
