+++ decidim-verifications/app/controllers/decidim/verifications/authorizations_controller.rb
@@ -6,9 +6,10 @@ module Decidim
     # shouldn't be necessary to expand it to add new authorization schemes.
     class AuthorizationsController < ApplicationController
       helper_method :handler, :unauthorized_methods
-      before_action :valid_handler, only: [:new, :create, :renew]
+      before_action :valid_handler, only: [:new, :create]
 
       include Decidim::UserProfile
+      include Decidim::Verifications::Renewable
       helper Decidim::DecidimFormHelper
       helper Decidim::CtaButtonHelper
       helper Decidim::AuthorizationFormHelper
@@ -42,25 +43,6 @@ module Decidim
           on(:invalid) do
             flash[:alert] = t("authorizations.create.error", scope: "decidim.verifications")
             render action: :new
-          end
-        end
-      end
-
-      def renew
-        authorization = Decidim::Authorization.find_by(
-          user: current_user,
-          name: handler.handler_name
-        )
-
-        DestroyUserAuthorization.call(authorization) do
-          on(:ok, authorization) do
-            flash[:notice] = t("authorizations.destroy.success", scope: "decidim.verifications")
-            redirect_to new_authorization_path(handler: authorization.name)
-          end
-
-          on(:invalid) do
-            flash[:alert] = t("authorizations.destroy.error", scope: "decidim.verifications")
-            redirect_to authorizations_path
           end
         end
       end
@@ -110,6 +92,15 @@ module Decidim
         return if redirect_url.blank? || !request.format.html?
 
         store_location_for(:user, redirect_url)
+      end
+
+      private
+
+      def authorization
+        @authorization ||= Decidim::Authorization.find_by(
+          user: current_user,
+          name: handler.handler_name
+        )
       end
     end
   end
