+++ decidim-verifications/app/controllers/decidim/verifications/csv_census/authorizations_controller.rb
@@ -4,6 +4,8 @@ module Decidim
   module Verifications
     module CsvCensus
       class AuthorizationsController < Decidim::ApplicationController
+        include Decidim::Verifications::Renewable
+
         helper_method :authorization
 
         before_action :load_authorization
