+++ decidim-verifications/app/controllers/decidim/verifications/sms/authorizations_controller.rb
@@ -4,6 +4,8 @@ module Decidim
   module Verifications
     module Sms
       class AuthorizationsController < ApplicationController
+        include Decidim::Verifications::Renewable
+
         helper_method :authorization
 
         def new
