+++ decidim-verifications/app/views/decidim/verifications/authorizations/_granted_authorization.html.erb
@@ -13,8 +13,6 @@
           <%= t("expires_at", scope: "decidim.authorization_handlers", timestamp: l(authorization.expires_at, format: :long)) %>
         <% end %>
       <% end %>
-      <br>
-      <%= authorization.wait_time %>
     </span>
   </div>
 </div>
