+++ decidim-verifications/lib/decidim/verifications/csv_census/engine.rb
@@ -10,7 +10,9 @@ module Decidim
         paths["lib/tasks"] = nil
 
         routes do
-          resource :authorization, only: [:new], as: :authorization
+          resource :authorization, only: [:new], as: :authorization do
+            get :renew, on: :collection
+          end
           root to: "authorizations#new"
         end
       end
