+++ decidim-verifications/lib/decidim/verifications/postal_letter/engine.rb
@@ -11,7 +11,9 @@ module Decidim
         paths["lib/tasks"] = nil
 
         routes do
-          resource :authorizations, only: [:new, :create, :edit, :update], as: :authorization
+          resource :authorizations, only: [:new, :create, :edit, :update], as: :authorization do
+            get :renew, on: :collection
+          end
 
           root to: "authorizations#new"
         end
